Refactored code to make it more object oriented.
