+++ src/test/java/uk/gov/companieshouse/accounts/association/models/email/senders/InvitationCancelledEmailSenderTest.java
@@ -1,0 +1,59 @@
+package uk.gov.companieshouse.accounts.association.models.email.senders;
+
+import static org.mockito.ArgumentMatchers.any;
+import static uk.gov.companieshouse.accounts.association.utils.Constants.INVITATION_CANCELLED_MESSAGE_TYPE;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.companieshouse.accounts.association.models.email.data.InvitationCancelledEmailData;
+import uk.gov.companieshouse.email_producer.EmailProducer;
+import uk.gov.companieshouse.email_producer.EmailSendingException;
+
+@ExtendWith(MockitoExtension.class)
+@Tag("unit-test")
+public class InvitationCancelledEmailSenderTest {
+
+    @Mock
+    EmailProducer emailProducer;
+
+    InvitationCancelledEmailSender invitationCancelledEmailSender;
+
+    @BeforeEach
+    public void setup(){
+        invitationCancelledEmailSender = new InvitationCancelledEmailSender( emailProducer );
+    }
+
+    @Test
+    void sendInvitationCancelledEmailWithNullInputsThrowsNullPointerException(){
+        Assertions.assertThrows( NullPointerException.class, () -> invitationCancelledEmailSender.sendEmail( "Krishna Patel", "Elon Musk", "Tesla" ).accept( null ) );
+        Assertions.assertThrows( NullPointerException.class, () -> invitationCancelledEmailSender.sendEmail(  null, "Elon Musk", "Tesla" ).accept( "kpatel@companieshouse.gov.uk" ) );
+        Assertions.assertThrows( NullPointerException.class, () -> invitationCancelledEmailSender.sendEmail(  "Krishna Patel", null, "Tesla" ).accept( "kpatel@companieshouse.gov.uk" ) );
+        Assertions.assertThrows( NullPointerException.class, () -> invitationCancelledEmailSender.sendEmail( "Krishna Patel", "Elon Musk", null ).accept( "kpatel@companieshouse.gov.uk" ) );
+    }
+
+    @Test
+    void sendInvitationCancelledEmailWithUnexpectedIssueThrowsBadRequestRuntimeException(){
+        Mockito.doThrow( new EmailSendingException( "Failed to send email", new Exception() ) ).when( emailProducer ).sendEmail( any(), any() );
+        Assertions.assertThrows( EmailSendingException.class, () -> invitationCancelledEmailSender.sendEmail( "Krishna Patel", "Elon Musk", "Tesla" ).accept( "kpatel@companieshouse.gov.uk" ) );
+    }
+
+    @Test
+    void sendInvitationCancelledEmailThrowsMessageOnKafkaQueue(){
+        final var emailData = new InvitationCancelledEmailData();
+        emailData.setTo( "kpatel@companieshouse.gov.uk" );
+        emailData.setSubject( "Companies House: Invitation cancelled for Elon Musk to be authorised to file online for Tesla" );
+        emailData.setPersonWhoCancelledInvite( "Krishna Patel" );
+        emailData.setPersonWhoWasCancelled( "Elon Musk" );
+        emailData.setCompanyName( "Tesla" );
+
+        invitationCancelledEmailSender.sendEmail( "Krishna Patel", "Elon Musk", "Tesla" ).accept( "kpatel@companieshouse.gov.uk" );
+        Mockito.verify( emailProducer ).sendEmail( emailData, INVITATION_CANCELLED_MESSAGE_TYPE );
+    }
+
+}
