+++ src/test/java/uk/gov/companieshouse/accounts/association/models/email/senders/InvitationRejectedEmailSenderTest.java
@@ -1,0 +1,57 @@
+package uk.gov.companieshouse.accounts.association.models.email.senders;
+
+import static org.mockito.ArgumentMatchers.any;
+import static uk.gov.companieshouse.accounts.association.utils.Constants.INVITATION_REJECTED_MESSAGE_TYPE;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.companieshouse.accounts.association.models.email.data.InvitationRejectedEmailData;
+import uk.gov.companieshouse.email_producer.EmailProducer;
+import uk.gov.companieshouse.email_producer.EmailSendingException;
+
+@ExtendWith(MockitoExtension.class)
+@Tag("unit-test")
+public class InvitationRejectedEmailSenderTest {
+
+    @Mock
+    EmailProducer emailProducer;
+
+    InvitationRejectedEmailSender invitationRejectedEmailSender;
+
+    @BeforeEach
+    public void setup(){
+        invitationRejectedEmailSender = new InvitationRejectedEmailSender( emailProducer );
+    }
+
+    @Test
+    void sendInvitationRejectedEmailWithNullInputsThrowsNullPointerException(){
+        Assertions.assertThrows( NullPointerException.class, () -> invitationRejectedEmailSender.sendEmail( "Elon Musk", "Tesla" ).accept( null ) );
+        Assertions.assertThrows( NullPointerException.class, () -> invitationRejectedEmailSender.sendEmail( (String) null, "Tesla" ).accept( "kpatel@companieshouse.gov.uk" ) );
+        Assertions.assertThrows( NullPointerException.class, () -> invitationRejectedEmailSender.sendEmail(  "Elon Musk", null ).accept( "kpatel@companieshouse.gov.uk" ) );
+    }
+
+    @Test
+    void sendInvitationRejectedEmailWithUnexpectedIssueThrowsBadRequestRuntimeException(){
+        Mockito.doThrow( new EmailSendingException( "Failed to send email", new Exception() ) ).when( emailProducer ).sendEmail( any(), any() );
+        Assertions.assertThrows( EmailSendingException.class, () -> invitationRejectedEmailSender.sendEmail( "Elon Musk", "Tesla" ).accept( "kpatel@companieshouse.gov.uk" ) );
+    }
+
+    @Test
+    void sendInvitationRejectedEmailThrowsMessageOnKafkaQueue(){
+        final var emailData = new InvitationRejectedEmailData();
+        emailData.setTo( "kpatel@companieshouse.gov.uk" );
+        emailData.setSubject( "Companies House: Elon Musk has declined to be digitally authorised to file online for Tesla" );
+        emailData.setPersonWhoDeclined( "Elon Musk" );
+        emailData.setCompanyName( "Tesla" );
+
+        invitationRejectedEmailSender.sendEmail( "Elon Musk", "Tesla" ).accept( "kpatel@companieshouse.gov.uk" );
+        Mockito.verify( emailProducer ).sendEmail( emailData, INVITATION_REJECTED_MESSAGE_TYPE );
+    }
+
+}
